requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
drf-yasg==1.21.4
python-decouple==3.6
gunicorn==20.1.0
whitenoise==6.2.0
pytest==7.1.3
pytest-django==4.5.2
coverage==6.4.4
django-filter==22.1